fix: return int -> size_t

+++ 11-binary_tree_size.c
@@ -7,7 +7,7 @@
  */
 size_t binary_tree_size(const binary_tree_t *tree)
 {
-	int left, right;
+	size_t left, right;
 
 	if (!tree)
 		return (0);
